requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
